Refactor computation registry stub into reusable module.
